Looks like enrolling is working.

+++ src/main/java/io/github/profilr/web/resources/PageEditTest.java
@@ -1,14 +1,13 @@
 package io.github.profilr.web.resources;
 
-import java.util.Map;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
-import java.util.ArrayList;
+import java.util.Map;
 
 import javax.inject.Inject;
 import javax.persistence.EntityManager;
 import javax.ws.rs.Consumes;
-
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
@@ -21,16 +20,16 @@ import javax.ws.rs.core.UriInfo;
 
 import org.glassfish.jersey.server.mvc.Template;
 
-import io.github.profilr.domain.Topic;
 import io.github.profilr.domain.Course;
 import io.github.profilr.domain.Question;
 import io.github.profilr.domain.Test;
-import io.github.profilr.web.PreAuth;
+import io.github.profilr.domain.Topic;
+import io.github.profilr.domain.User;
 import io.github.profilr.web.Session;
+import io.github.profilr.web.UserNotAuthorizedException;
 import io.github.profilr.web.WebResource;
 
 @Path("edit-test/{test-id}")
-@PreAuth //TODO REMOVE ME ASAP - ONLY FOR TESTING
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
 public class PageEditTest extends WebResource {
@@ -48,8 +47,13 @@ public class PageEditTest extends WebResource {
 	@GET
 	@Template(name="/edittest")
 	@Produces(MediaType.TEXT_HTML)
-	public Response get() {
+	public Response get() throws UserNotAuthorizedException {
 		Test t = entityManager.find(Test.class, testID);
+		
+		User u = (User) session.get("user");
+		if (!u.isCourseAdmin(t.getCourse()))
+			throw new UserNotAuthorizedException();
+		
 		return Response.ok(getView(t)).build();
 	}
 	
@@ -78,8 +82,14 @@ public class PageEditTest extends WebResource {
 	
 	@POST
 	@Path("create-question")
-	public Response createQuestion(Question question) {
-		question.setTest(entityManager.find(Test.class, testID));
+	public Response createQuestion(Question question) throws UserNotAuthorizedException {
+		Test t = entityManager.find(Test.class, testID);
+		
+		User u = (User) session.get("user");
+		if (!u.isCourseAdmin(t.getCourse()))
+			throw new UserNotAuthorizedException();
+		
+		question.setTest(t);
 		entityManager.persist(question);
 		return Response.created(uriInfo.getBaseUriBuilder()
 									   .path(PageEditTest.class, "editQuestion")
@@ -90,17 +100,28 @@ public class PageEditTest extends WebResource {
 	
 	@POST
 	@Path("edit-question/{question-id}")
-	public Response editQuestion(Question question, @PathParam("question-id") int questionID) {
+	public Response editQuestion(Question question, @PathParam("question-id") int questionID) throws UserNotAuthorizedException {
+		Test t = entityManager.find(Test.class, testID);
+		
+		User u = (User) session.get("user");
+		if (!u.isCourseAdmin(t.getCourse()))
+			throw new UserNotAuthorizedException();
+		
 		question.setQuestionID(questionID);
-		question.setTest(entityManager.find(Test.class, testID));
+		question.setTest(t);
 		entityManager.merge(question);
 		return Response.ok().build();
 	}
 	
 	@POST
 	@Path("delete-question/{question-id}")
-	public Response deleteQuestion(@PathParam("question-id") int questionID) {
+	public Response deleteQuestion(@PathParam("question-id") int questionID) throws UserNotAuthorizedException {
 		Question q = entityManager.find(Question.class, questionID);
+		
+		User u = (User) session.get("user");
+		if (!u.isCourseAdmin(q.getTest().getCourse()))
+			throw new UserNotAuthorizedException();
+		
 		entityManager.remove(q);
 		return Response.ok().build();
 	}
